% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/about.R
\name{about}
\alias{about}
\title{About description module}
\usage{
about(id)
}
\arguments{
\item{id}{\code{character} ID of this module}
}
\description{
About description module
}
\author{
Sebastian Wolf \email{sebastian@mail-wolf.de}
}
