% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_html.R
\name{attachstravavisDep}
\alias{attachstravavisDep}
\title{Attach Stylesheet dependencies and js dependencies}
\usage{
attachstravavisDep(tag, dependency = "style.css",
  dependency_script = NULL)
}
\arguments{
\item{tag}{a shiny element that the dependencies shall be added to}

\item{dependency}{\code{character} A list of css files located in inst/www of the package}

\item{dependency_script}{\code{character} A list of js files located in inst/www of the package}
}
\description{
Attach Stylesheet dependencies and js dependencies
}
\details{
Taken from shinyWidgets package
}
