% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{barChartOutput}
\alias{barChartOutput}
\title{Shiny output for a two sided barchart}
\usage{
barChartOutput(id, label = "", color_left = "#fc4c02",
  color_right = "#777777", height = "3em")
}
\arguments{
\item{id}{ID}

\item{label}{Label in the center of the barchart}

\item{color_left}{HEX color code or HTML rgb color code of the left side of the chart}

\item{color_right}{HEX color code or HTML rgb color code of the right side of the chart}

\item{height}{HTML height string in px or em}
}
\description{
This is linked to the 'barchart-binding.js' and is generated with the 'simple-skillbar.js'
}
