% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_distance_city_stravasegments}
\alias{calc_distance_city_stravasegments}
\title{Function returning the lines of strava segments within a city radius}
\usage{
calc_distance_city_stravasegments(city = "London", radius = 3000)
}
\arguments{
\item{city}{Name of the city}

\item{radius}{Radius in m}
}
\description{
Function returning the lines of strava segments within a city radius
}
\details{
To use this function you need to call \code{data(all_data_table_strava)}
}
