% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_score}
\alias{calc_score}
\title{Calculate the score for a given segment}
\usage{
calc_score(statistics = "med", elevation_factor = 1,
  city_data_sets = data.frame(median = c(), average = c(), climb = c()))
}
\arguments{
\item{statistics}{\code{med} or \code{avg}}

\item{elevation_factor}{\code{numeric} between 0 and 5}

\item{city_data_sets}{\code{data.frame} DF containing columns 'median', 'average' and 'climb'}
}
\value{
\code{city_data_sets} with an an additional column 'score'
}
\description{
SCORE is calculated as MEDIAN + ELEVATION_STATS * 0.5 * CLIMB
OR AVERAGE + ELEVATION_STATS * 0.5 * CLIMB
}
\details{
CLIMB shall represent the average slope of the segment
}
\author{
Sebastian Wolf
}
