% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city.R
\name{city}
\alias{city}
\title{city Module server function}
\usage{
city(input, output, session)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session}
}
\value{
The data for all segments of the city
}
\description{
city Module server function
}
