% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_count.R
\name{city_count}
\alias{city_count}
\title{Server function of city_count module}
\usage{
city_count(input, output, session, city_data = NULL)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session}

\item{city_data}{\code{reactive} representing strava crawled data from at
least two different cities}
}
\description{
This module shows how many segments can be seen for each city
It contains a plot for the average length of a segment.
Additionally it contains the \link{city_hist} module twice
to show the histogram of the segment length for each city.
}
