% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_hist.R
\name{city_hist}
\alias{city_hist}
\title{City Histogram length of segments module}
\usage{
city_hist(input, output, session, which_id = 1, city_data = NULL)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session}

\item{which_id}{\code{numeric} 1 or 2}

\item{city_data}{\code{reactive - data.frame} reactive giving a data frame
with a distance column and a city column}
}
\value{
Set the \code{output$segment_hist} part of the module to
   contain a ggplot histogram of the segments binned by length
}
\description{
City Histogram length of segments module
}
