% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_score.R
\name{city_score}
\alias{city_score}
\title{city_score module}
\usage{
city_score(input, output, session, city_data = NULL, stats = NULL)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session 

uses \link{calc_score} to calculate the score for each segment}

\item{city_data}{\code{reactive} representing strava crawled data from at
least two different cities}

\item{stats}{\code{reactive} A reactive element representing the outcome
of the \link{filtering} module that contains the filters to be used for
statistics.}
}
\description{
city_score module
}
\author{
Sebastian Wolf
}
