% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_score_hist.R
\name{city_score_hist}
\alias{city_score_hist}
\title{City Score Histogram Module}
\usage{
city_score_hist(input, output, session, city_data = NULL, stats = NULL)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session}

\item{city_data}{\code{reactive} representing strava crawled data from at
least two different cities}

\item{stats}{\code{reactive} A reactive element representing the outcome
of the \link{filtering} module that contains the filters to be used for
statistics.}
}
\description{
Create a left right plot for the scores of segments binned by the
length of the segments
}
\author{
Sebastian Wolf \email{sebastian@mail-wolf.de}
Plot left/right statistics
}
