% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_statistic.R
\name{city_statistic}
\alias{city_statistic}
\title{Module server function}
\usage{
city_statistic(filters = NULL, city_data = NULL)
}
\arguments{
\item{filters}{list of filters}

\item{city_data}{Data frame containing segments for two cities}
}
\value{
data filtered by statistic
}
\description{
Module server function
}
