% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_count.R
\name{cityCountUI}
\alias{cityCountUI}
\title{UI function to generate a segment overview per city}
\usage{
cityCountUI(id)
}
\arguments{
\item{id}{ID}
}
\description{
UI function to generate a segment overview per city
}
