% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_score_hist.R
\name{cityhistScoreUI}
\alias{cityhistScoreUI}
\title{City Score Histogram UI Module}
\usage{
cityhistScoreUI(id)
}
\arguments{
\item{id}{\code{character} ID of this module}
}
\description{
City Score Histogram UI Module
}
\author{
Sebastian Wolf \email{sebastian@mail-wolf.de}
}
