% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_score.R
\name{cityScoreUI}
\alias{cityScoreUI}
\title{Score output - Function providing a shiny Output}
\usage{
cityScoreUI(id)
}
\arguments{
\item{id}{\code{character} ID of the element}
}
\description{
This function generates 1 plot giving the overall score
of two cities. Additionally it can integrate a \link{city_score_hist}
}
\author{
Sebastian Wolf
}
