% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city.R
\name{citySelect}
\alias{citySelect}
\title{City Selection Module
UI Function to select a city + Range}
\usage{
citySelect(id, selected = "London", selections = c(London = "London",
  Paris = "Paris", Berlin = "Berlin"))
}
\arguments{
\item{id}{\code{character} ID of the element}

\item{selected}{\code{character} Name of the city that shall be selected}

\item{selections}{\code{character}  named vector to allow cities to go into
the select input}
}
\description{
City Selection Module
UI Function to select a city + Range
}
\author{
Sebastian Wolf
}
