% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filtering}
\alias{filtering}
\title{Compile a list of filters from such inputs - shiny module}
\usage{
filtering(input, output, session)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\value{
filters \code{reactive} list of filters
}
\description{
Compile a list of filters from such inputs - shiny module
}
\author{
Sebastian Wolf
}
