% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filteringUI}
\alias{filteringUI}
\title{Function to select a the filters of the statistics}
\usage{
filteringUI(id)
}
\arguments{
\item{id}{\code{character} ID of the object inside shiny app}
}
\description{
This function creates an output that contains a
\itemize{
\item{gender}{ male, femal, both}
\item{distance}{ length of the segment}
\item{elevation}{ 0-5 = Elevation score}
\item{chaser}{ # of participants in Top30}
\item{score}{ Whether the statistics are calculated by means or medians}
}
}
\author{
Sebastian Wolf
}
