% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{renderBarChart}
\alias{renderBarChart}
\title{Generic function to render the barchart}
\usage{
renderBarChart(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{shall give a list with elements left, right and unit}

\item{env}{environment}

\item{quoted}{leave FALSE}
}
\description{
Generic function to render the barchart
}
