% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_text.R
\name{score_text}
\alias{score_text}
\title{Module server function}
\usage{
score_text(input, output, session, scores = NULL, city_data = NULL)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session}

\item{scores}{scoring table to calculate which city won}

\item{city_data}{city data to extract the city_names from}
}
\value{
set the \code{output$score_text_out} of this module
 to 'XXX is the winner' while replacing XX with the city
 with the higher score in the city_data
}
\description{
Module server function
}
