% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_text.R
\name{scoreTextUI}
\alias{scoreTextUI}
\title{Function to select a city + Range}
\usage{
scoreTextUI(id)
}
\arguments{
\item{id}{ID}
}
\description{
Function to select a city + Range
}
\author{
Sebastian Wolf
}
