% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{segment_map}
\alias{segment_map}
\title{Create a Leaflet map with radius + circle markers}
\usage{
segment_map(segment_data = data.frame(id = character(0), lat = 0, long =
  0, name = character(0)), radius = 2000, marker_list = data.frame(lon
  = 0, lat = 45), zoom = 4)
}
\arguments{
\item{segment_data}{\code{data.frame} DF with cols lng, lat, name, id
for each segment}

\item{radius}{\code{numeric} in meter}

\item{marker_list}{\code{numeric} center of the map containing lon and lat variable}

\item{zoom}{\code{numeric} leaflet zoom}
}
\description{
Create a map of a certain location. It is surrounded by a ring
in a certain radius. Additionally for each segment of \code{segment_data}
a point will be created with the link to the STRAVA segment
}
\author{
Sebastian Wolf
}
