% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{server}
\alias{server}
\title{Shiny server function of stravavis}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session}
}
\description{
Shiny server function of stravavis
}
\author{
Sebastian Wolf \email{sebastian@mail-wolf.de}
}
