% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/share.R
\name{share}
\alias{share}
\title{UI to create a share link from filters}
\usage{
share(input, output, session, filters)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session}

\item{filters}{reactive giving a list of filters used to modify the data}

\item{id}{character value of ID of this UI}
}
\description{
UI to create a share link from filters
}
