% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{ui}
\alias{ui}
\title{UI function of the stravachaser app}
\usage{
ui()
}
\description{
Function visualizing a shiny app for strava city comparison
}
\author{
Sebastian Wolf \email{sebastian@mail-wolf.de}
}
